use modal componet and seperate display logic out
